Seperate training utils and utils
